requirements.txt
streamlit>=1.30.0
requests>=2.31.0
beautifulsoup4>=4.12.0
gspread>=6.0.0
google-auth>=2.25.0
pandas>=2.1.0
